Changes to the Grid and its implementation the MutableGrid class
- Mutable grid now inner class of GameController
- `boolean markIs(Point point, Mark markType)` is now `Mark getMark(Point point)`
- added some additional documentation in Grid and MutableGrid
